Move lifecycle from pkg to model
