+++ cmd/swift.go
@@ -7,9 +7,9 @@ import (
 	"os"
 	"strings"
 
+	"github.com/cozy/cozy-stack/model/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/config/config"
 	"github.com/cozy/cozy-stack/pkg/instance"
-	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/swift"
 	"github.com/spf13/cobra"
 )
