+++ model/app/installer_konnector_test.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/cozy/afero"
 	"github.com/cozy/cozy-stack/model/app"
+	"github.com/cozy/cozy-stack/model/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/config/config"
 	"github.com/cozy/cozy-stack/pkg/consts"
-	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/stretchr/testify/assert"
 )
