+++ web/auth/rate_limiting.go
@@ -3,8 +3,8 @@ package auth
 import (
 	"fmt"
 
+	"github.com/cozy/cozy-stack/model/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/instance"
-	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/limits"
 )
