+++ web/routing.go
@@ -6,9 +6,9 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/cozy/cozy-stack/model/instance/lifecycle"
 	build "github.com/cozy/cozy-stack/pkg/config"
 	"github.com/cozy/cozy-stack/pkg/config/config"
-	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/jsonapi"
 	"github.com/cozy/cozy-stack/pkg/metrics"
 	"github.com/cozy/cozy-stack/web/accounts"
